Optimized CI for Firebase deployment
Simplified and improved the CI workflow for Flutter by directly using
Firebase tools for deployment. Removed the previously used action for
APK upload and artifact handling, and instead switched to a streamlined
process that installs Firebase CLI globally, handles authentication, and
directly deploys to Firebase App Distribution. This change reduces the
complexity of the workflow and leverages Firebase's native capabilities
for increased reliability and maintainability.