Updated CI with artifact debug and Firebase action
Changed the GitHub Artifact action to use a tagged version for enhanced
stability. Added a step to display the structure of downloaded files to
aid in debugging. Switched to a more recent version of the Firebase
Distribution GitHub Action without hardcoded release notes, allowing for
dynamic control over release messaging in future.