Introduced Custom Color Theming
Implemented a defined color system for the Flutter application that
includes both a new enumeration for app-specific colors and the addition
of custom MaterialColor definitions to standardize the UI design across
the application. The update introduces a set of primary and accent
colors, laying the foundation for a consistent theming approach and
improved visual aesthetics. This change makes color usage across the
application more maintainable and enables easier theming adjustments in
the future.