Updated Flutter Release APK Build and App Version to 0.0.5(5)
Adjusted the Flutter CI workflow to include the `--no-tree-shake-icons`
option during the APK build process, ensuring that all icons are
preserved in the release build. This change resolves potential issues
with missing icons that may appear after the tree shaking optimization.
Additionally, the app version in `pubspec.yaml` has been incremented,
signifying a new iteration of the application in preparation for the
next release cycle.